[DEV-376] Changed any instance of %v in format strings with a more specific format token

+++ util/amount.go
@@ -85,7 +85,7 @@ func NewAmount(f float64) (Amount, error) {
 		return 0, errors.New("invalid bitcoin amount")
 	}
 
-	return round(f * SatoshiPerBitcoin), nil
+	return round(f * float64(SatoshiPerBitcoin)), nil
 }
 
 // ToUnit converts a monetary amount counted in bitcoin base units to a
